dpif-netdev: change execute->flow->in_port into odp_port
the flow->in_port value is previously ignored. We normally use
the the md->in_port of the first packet in the batch to get the
value. However, since some actions might need to be redo due to
the fix of ipf issues, we might have an empty batch to feed the
action execute, so the conntrack makes use of flow->in_port
to get the in_port value and form the ipf_ctx.

This patch changes execute->flow->in_port into odp_port before
doing action executes. with this patch, the ipf_ctx patch will
not cause testsuite report errors.

Signed-off-by: Peng He <hepeng.0320@bytedance.com>